requirements.txt
google-generativeai
groq
requests
beautifulsoup4
PyPDF2
python-dotenv
termcolor
ollama
anthropic
markdown
xhtml2pdf
